requirements.txt
autopep8==1.5.2
bcrypt==3.1.7
Cerberus==1.3.2
cffi==1.14.0
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
flask-mongoengine==0.9.5
Flask-PyMongo==2.3.0
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.20.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pymongo==3.10.1
python-dotenv==0.13.0
six==1.15.0
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
